delete-data: move streamVideo functions to delete-data
